% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latex.R
\name{add_latex}
\alias{add_latex}
\title{Quick paste a LaTex object}
\usage{
add_latex(type = "bmatrix")
}
\arguments{
\item{type}{The type of equations in Latex to choose.}
}
\description{
Quick paste a LaTex object
}
\author{
Jiaxiang Li
}
